Update: move stuff, Add: integrate new config, Fix: crash on empty presets
